[ test ] Update tests

+++ test/org/ice1000/julia/lang/execution/julia-execution-test.kt
@@ -7,52 +7,51 @@ import org.ice1000.julia.lang.shouldBe
 import org.junit.Test
 import java.util.concurrent.TimeUnit
 
-class JuliaExecutionTest {
-	@Test
-	fun test() {
-		if (!System.getenv("CI").isNullOrBlank()) return
-		val process = Runtime.getRuntime().exec("/home/ice1000/SDK/julia-6.2/bin/julia -q").also {
-			//language=Julia
-			it.outputStream.let {
-				it.write("using DocumentFormat: format\n".toByteArray())
-				it.flush()
-			}
-		}
-		println("0")
+fun main(args: Array<String>) {
+	val process = Runtime.getRuntime().exec("/home/ice1000/SDK/julia-6.2/bin/julia -q").also {
 		//language=Julia
-		process.outputStream.let {
-			it.write("println(\"Bye, world!\")\n".toByteArray())
+		it.outputStream.let {
+			it.write("using DocumentFormat: format\n".toByteArray())
 			it.flush()
 		}
-		process.waitFor(5, TimeUnit.SECONDS)
-		println("1")
-		val reader = process.inputStream.bufferedReader()
-		reader.readLine().let(::println)
-		println("2")
-		//language=Julia
-		process.outputStream.let {
-			it.write("println(\"Bye, world again!\")\n".toByteArray())
-			it.flush()
-		}
-		process.waitFor(5, TimeUnit.SECONDS)
-		println("1")
-		reader.readText().let(::println)
-		println("2")
 	}
+	println("0")
+	//language=Julia
+	process.outputStream.let {
+		it.write("println(\"Bye, world!\")\n".toByteArray())
+		it.flush()
+	}
+	process.waitFor(5, TimeUnit.SECONDS)
+	println("1")
+	val reader = process.inputStream.bufferedReader()
+	reader.readLine().let(::println)
+	println("2")
+	//language=Julia
+	process.outputStream.let {
+		it.write("println(\"Bye, world again!\")\n".toByteArray())
+		it.flush()
+	}
+	process.waitFor(5, TimeUnit.SECONDS)
+	println("1")
+	reader.readText().let(::println)
+	println("2")
+}
 
+class JuliaExecutionTest {
 	@Test
 	fun testDocker() {
 		//Windows
 		if (!System.getenv("CI").isNullOrBlank()) return
+		if (!SystemInfo.isWindows) return
 		val winCD = "%CD%"
 		val unixPWD = "\$PWD"
 		val currentDir = if (SystemInfo.isWindows) winCD else unixPWD
 		//windows command need not `-it`
 		//currentDir in @Test is Project Root
-		val juliaScriptFile="ParseFunctions.jl"
-		val params=""
+		val juliaScriptFile = "ParseFunctions.jl"
+		val params = ""
 		val cmd = "cmd /c docker run --rm -v \"$currentDir/testData\":/usr/myapp -w /usr/myapp julia julia $juliaScriptFile $params"
-		val process=Runtime.getRuntime().exec(cmd)
+		val process = Runtime.getRuntime().exec(cmd)
 		println(process.inputStream.reader().readText())
 		println("----error----")
 		println(process.errorStream.reader().readText())
@@ -66,7 +65,6 @@ class JuliaExecutionTest {
 class JuliaConfig {
 	@Test
 	fun testFile() {
-		if (!System.getenv("CI").isNullOrBlank()) return
 		println(FileUtilRt.getNameWithoutExtension("/home/zh/IdeaProjects/JuliaTest/src/lk.jl"))
 		println(Files.getNameWithoutExtension("/home/zh/IdeaProjects/JuliaTest/src/lk.jl"))
 	}
@@ -81,6 +79,7 @@ class JuliaConsoleTest {
 	@Test
 	fun testShouldFolding() {
 		if (!System.getenv("CI").isNullOrBlank()) return
+		if (!SystemInfo.isMac) return
 		val command = "/Applications/Julia-0.6.app/Contents/Resources/julia/bin/julia --check-bounds=no --history-file=no --inline=no --color=no --math-mode=ieee --handle-signals=no --startup-file=no --optimize=0 --compile=yes -q /Users/paul/IdeaProjects/julia-project-test2/src/Jul.jl 23 f jh"
 		JuliaConsoleFolding().shouldFoldLine(command) shouldBe true
 	}
